Upgrade yaml pkg (#100)
Switches to a better maintained version of `yaml` which switches over to `yaml.v3`. This was needed because `yaml.v2` (used by the prior `yaml` pkg) pushes long strings over on to a newline which can cause issues with some YAML parses (we at Airbnb do some pre-processing of YAML files before loading them in to Prometheus/Mimir).